backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
pypdf
chromadb
python-dotenv
pydanticpsycopg2-binary
